added the methods for operator checks in ReportingHelper class and moved the rsi step to common step class
