+++ tests/acceptance-tests/features/rsi/rsi_validation_qvqt.feature
@@ -20,7 +20,7 @@ Feature: RSI Survey - QvQT validation rule
       | reference   | period | values    | totalTurnoverValue | deriveValue | result | operator     | thresholdValue | validation                                                | isValidationExists |
       | 49900694171 | 201903 | 1,1,1,0,1 | 10                 | 4           | true   | greater than | 5              | Total different to calculated total (allowing for margin) | be                 |
       | 49900694171 | 201903 | 1,2,3,2,2 | 4                  | 10          | true   | greater than | 5              | Total different to calculated total (allowing for margin) | be                 |
-      | 49900694171 | 201903 | 1,1,1,1,1 | 10                 | 5           | false  | equal than   | 5              | Total different to calculated total (allowing for margin) | not be             |
-      | 49900694171 | 201903 | 1,2,3,2,2 | 5                  | 10          | false  | equal than   | 5              | Total different to calculated total (allowing for margin) | not be             |
+      | 49900694171 | 201903 | 1,1,1,1,1 | 10                 | 5           | false  | equal to     | 5              | Total different to calculated total (allowing for margin) | not be             |
+      | 49900694171 | 201903 | 1,2,3,2,2 | 5                  | 10          | false  | equal to     | 5              | Total different to calculated total (allowing for margin) | not be             |
       | 49900694171 | 201903 | 1,1,1,1,2 | 10                 | 6           | false  | less than    | 5              | Total different to calculated total (allowing for margin) | not be             |
       | 49900694171 | 201903 | 1,2,3,2,2 | 6                  | 10          | false  | less than    | 5              | Total different to calculated total (allowing for margin) | not be             |
